src/requirements.txt
Flask==2.3.2
gunicorn==20.1.0
SQLAlchemy==2.0.17
google-cloud-storage==2.10.0
pg8000==1.29.8